# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Warriors Gym Membership API
  version: 1.0.0
  description: |
    REST API behind the Warriors Gym admin web app, admin mobile app and member mobile app.

    **Authentication.** Sign in at `/auth/sign-in/{userType}`, then send the returned token as
    `Authorization: Bearer <token>`. Admin routes need an admin token and member routes a member
    token; a token for one role is rejected on the other.

    **Errors.** Every error is `{ "error": "message" }` with a meaningful status
    (400 validation, 401 auth, 403 forbidden, 404 missing, 409 conflict, 413 too large, 429 rate limited).

    **Demo.** When the server runs in demo mode, sample data is seeded and the two published demo
    logins cannot be changed. Data resets nightly.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
...
  /member/me:
    get:
      tags: [Member]
      summary: The signed-in member
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Member profile
          content:
            application/json:
              schema:
                type: object
                properties: { member: { $ref: "#/components/schemas/Member" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /member/profile:
    put:
      tags: [Member]
      summary: Update the signed-in member's profile
      description: Editable fields are full_name, gender, height, weight, age and language. Send `oldPassword` and `password` to change the password, or `phone_number` to change it; a phone change returns a fresh token.
      security: [{ bearerAuth: [] }]
      requestBody: { $ref: "#/components/requestBodies/ProfileForm" }
      responses:
        "200": { $ref: "#/components/responses/ProfileUpdated" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /member/membership_plans:
    get:
      tags: [Member]
      summary: Plans currently on sale
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Active plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  membershipPlan:
                    type: array
                    items: { $ref: "#/components/schemas/MembershipPlan" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /member/membership:
    get:
      tags: [Member]
      summary: The member's live membership with days and tickets left
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Membership
          content:
            application/json:
              schema:
                type: object
                properties:
                  membership:
                    allOf:
                      - $ref: "#/components/schemas/Membership"
                      - type: object
                        properties:
                          daysLeft: { type: integer }
                          remainingTicket: { type: integer, description: Ticket plans only }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /member/membership/off:
    post:
      tags: [Member]
      summary: Buy or renew a membership (simulated checkout)
      description: |
        Records a `Pending` payment at the plan's database price and activates the membership; no
        money moves until the Chapa integration is enabled. With `isNew: true`, `id` is a plan id;
        otherwise `id` is the member's own membership id to renew. Prices and durations in the
        request body are ignored.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string, format: uuid }
                isNew: { type: boolean }
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                type: object
                properties: { membership: { type: string, enum: [new success, renew success] } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /member/membershipRenew:
    put:
      tags: [Member]
      summary: Renew the member's own membership
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [membership_id]
              properties: { membership_id: { type: string, format: uuid } }
      responses:
        "200":
          description: Renewed
          content:
            application/json:
              schema:
                type: object
                properties: { membership: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /member/membershipCancel:
    put:
      tags: [Member]
      summary: Cancel the member's own membership
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties: { id: { type: string, format: uuid } }
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties: { membership: { type: string } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /member/transactions:
    get:
      tags: [Member]
      summary: The member's 30 most recent payments
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Payments, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items: { $ref: "#/components/schemas/Transaction" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /member/attendanceLog:
    get:
      tags: [Member]
      summary: The member's 20 most recent check-ins, oldest first
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Check-ins
          content:
            application/json:
              schema:
                type: object
                properties:
                  attendanceLog:
                    type: array
                    items: { $ref: "#/components/schemas/AttendanceLog" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /member/workoutPlan:
    get:
      tags: [Member]
      summary: All workouts grouped by type
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Workouts
          content:
            application/json:
              schema:
                type: object
                properties:
                  workoutPlan: { $ref: "#/components/schemas/WorkoutsByType" }
                  length: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /member/programs:
    get:
      tags: [Member]
      summary: The 50 newest programs
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: object
                properties:
                  programs:
                    type: array
                    items: { $ref: "#/components/schemas/Program" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  requestBodies:
    ProfileForm:
      required: true
      content:
        multipart/form-data:
          schema:
            type: object
            required: [metadata]
            properties:
              metadata: { type: string, description: JSON object with the fields to change }
              file: { type: string, format: binary, description: Optional profile image }
  responses:
    BadRequest:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid or expired token, or wrong role
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Not allowed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ProfileUpdated:
      description: Updated. A phone-number change also returns a new token and the user.
      content:
        application/json:
          schema:
            oneOf:
              - type: object
                properties: { user: { type: string, example: successfully updated } }
              - type: object
                properties:
                  token: { type: string }
                  user: { type: object }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
...
    Member:
      type: object
      properties:
        id: { type: string, example: "00042", description: Zero-padded }
        full_name: { type: string }
        phone_number: { type: string }
        gender: { type: string }
        height: { type: integer }
        weight: { type: integer }
        age: { type: integer }
        activity_status: { type: string, enum: [Active, Payment Due, Inactive] }
        language: { type: string }
        image: { type: string, nullable: true }
        registration_Date: { type: string }
...
    MembershipPlanInput:
      type: object
      properties:
        membership_name: { type: string }
        plan_type: { type: string, enum: [Daily, Ticket] }
        duration_days: { type: integer, minimum: 1 }
        fee: { type: integer, minimum: 0 }
        ticket_amount: { type: integer, minimum: 1, description: Required for Ticket plans }
        description: { type: string }
        status: { type: string, enum: [Active, Inactive] }
    MembershipPlan:
      allOf:
        - $ref: "#/components/schemas/MembershipPlanInput"
        - type: object
          properties:
            id: { type: string, format: uuid }
    Membership:
      type: object
      properties:
        id: { type: string, format: uuid }
        member_id: { type: integer }
        membership_plan_id: { type: string, format: uuid }
        start_date: { type: string, format: date-time }
        end_date: { type: string, format: date-time }
        ticket: { type: integer, nullable: true }
        status: { type: string, enum: [Active, Payment Due, Inactive, Cancled] }
        membershipPlan: { $ref: "#/components/schemas/MembershipPlan" }
    AttendanceLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        member_id: { type: integer }
        membership_id: { type: string, format: uuid }
        check_in: { type: string, format: date-time }
        check_out: { type: string, nullable: true }
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        payer_id: { type: integer }
        membershipPlan_id: { type: string, format: uuid }
        payment_method: { type: string }
        amount: { type: integer }
        payment_for: { type: string }
        paid_at: { type: string }
        status: { type: string, enum: [Pending, Successfull] }
        payer: { $ref: "#/components/schemas/Member" }
    Video:
      type: object
      properties:
        id: { type: string, format: uuid }
        path: { type: string, example: uploads/videos/1758000000000-uuid.mp4, description: Prefix with the server address to get a URL }
        size: { type: integer, description: Bytes on disk, 0 if unavailable }
    WorkoutPlan:
      type: object
      properties:
        id: { type: string, format: uuid }
        workout_title: { type: string }
        workout_type: { type: string }
        workout_level: { type: string }
        workout_rep: { type: integer }
        workout_sets: { type: integer }
        workout_break: { type: string }
        video: { $ref: "#/components/schemas/Video" }
    WorkoutsByType:
      type: object
      description: Workouts keyed by their `workout_type`
      additionalProperties:
        type: array
        items: { $ref: "#/components/schemas/WorkoutPlan" }
    Program:
      type: object
      properties:
        id: { type: string, format: uuid }
        content: { type: string }